Remove validate_configuration function
